Use saner number types, instead of float64 everywhere.

+++ server/events.go
@@ -2,18 +2,18 @@ package main
 
 type FireEvent struct {
 	PlayerId     string
-	Time         float64
+	Time         uint64
 	ProjectileId string
 }
 
 type ChangeAccelerationEvent struct {
 	PlayerId  string
-	Time      float64
-	Direction float64 // TODO: try to use an int8 here (and try to switch other types to better types instead of float64 as well)
+	Time      uint64
+	Direction int8 // TODO: try to use an int8 here (and try to switch other types to better types instead of float64 as well)
 }
 
 type ChangeRotationEvent struct {
 	PlayerId  string
-	Time      float64
-	Direction float64 // TODO: try to use an int8 here (and try to switch other types to better types instead of float64 as well)
+	Time      uint64
+	Direction int8 // TODO: try to use an int8 here (and try to switch other types to better types instead of float64 as well)
 }
